ai_engine/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
ultralytics>=8.0.0
opencv-python>=4.8.0
python-multipart>=0.0.6
requests>=2.31.0
pyyaml>=6.0.1
paddlepaddle
paddleocr==3.4.0
scipy>=1.11.1
numpy>=1.24.0
